[Test] Patch record test
